Added README (+ license)
I didn't really want a license at all, but apparently you cannot release
your work into the public domain under German law.

The closest thing to public domain is probably the WTFPL, but I don't like
that license, partly because of the somewhat immature wording.

I settled on the ISC license, because it is permissive, short and easy to
understand.