MENACE Bugfix: Use permutation to get canonical_grid legal moves
It took me hours to find this stupid oversight. I was using the legal
moves of the grid that replays the game given to end_of_game() instead of
using the permutation on that to get the legal moves of the canonical grid.

Fixed by rewriting the passage in question completely.
Learned to use #collect in the process. :-)