Grid: Remove a superflous check
I originally checked whether the field that was just changed is in a given
array before checking whether the row/column/diagonal represented by that
array holds three matching symbols.

I don't remember why I did that, because this requires an additional
parameter (changed_field) for #adjust_gamestate and is probably slower,
because we access the array twice: once for verifying the changed_field
is in there, and once for the actual comparison.