MENACE: Change learning algorithm
I remembered something about the MENACE algorithm I had forgotten before:
Instead of preventing MENACE from setting the weight of a move to zero,
Michie restarted MENACE's learning for a particular gamestate whenever it
had learned that all moves in that state are bad.

Compared to what I was doing before, this means that bad moves are weeded
out completely, instead of being made less frequently, which should speed up
"learning not to lose" a lot.